refactor: 컨트롤러에서 page 관련된 항목을 PagesController로 분리
