+++ src/main/java/roomescape/pages/PagesController.java
@@ -1,12 +1,17 @@
-package roomescape.home;
+package roomescape.pages;
 
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.GetMapping;
 
 @Controller
-public class HomeController {
+public class PagesController {
 	@GetMapping("/")
 	public String homePage() {
 		return "home";
+	}
+	
+	@GetMapping("/reservation")
+	public String reservationPage() {
+		return "reservation";
 	}
 }
